+++ src/config/types.ts
@@ -51,7 +51,6 @@ export type Options = {
   secrets: boolean;
   strict: boolean | undefined;
   ignoreUrls?: string[];
-  noCompare: boolean;
   uppercaseKeys: boolean;
   expireWarnings: boolean;
   inconsistentNamingWarnings: boolean;
@@ -83,7 +82,6 @@ export type RawOptions = {
   secrets?: boolean;
   strict?: boolean;
   ignoreUrls?: string[];
-  noCompare?: boolean;
   init?: boolean;
   uppercaseKeys?: boolean;
   expireWarnings?: boolean;
@@ -139,7 +137,6 @@ export interface ScanOptions {
   files?: string[];
   secrets?: boolean;
   ignoreUrls?: string[];
-  noCompare?: boolean;
 }
 
 export interface ScanResult {